GH-29 Make server usable in production Fixes #29
- Moved code to create server from index.js to server.js
- Update index.js to simply import the server function and invoke it,
  while exporting the result. This will allow us to still control the
  server in integration tests.
- Update index test to simply require the index file. There is no longer
  a need to invoke a function.
- Update integration test to simply require the bundle and assign the
  exported value to "server". This way, we can still stop the server
  when each test completes.
