+++ packages/srcmap-reverse/test/index.integration.test.js
@@ -1,5 +1,4 @@
 import { describe, beforeEach, afterEach, it, expect } from './jasmine.js';
-import startServer from '../dist/bundle.js';
 import getReq from '@iml/req';
 import fs from 'fs';
 
@@ -7,7 +6,7 @@ describe('srcmap-reverse integration test', () => {
   let trace, reversedFixture, server;
 
   beforeEach(() => {
-    server = startServer();
+    server = require('../dist/bundle.js');
     trace = fs.readFileSync(`${__dirname}/fixtures/trace.txt`, 'utf8');
 
     reversedFixture = fs
